Add $version to some queries
The intention was that this script could handle the situation where
an .iso with the same unversioned name but different contents existed in
two separate CentOS versions. Think boot.iso, although that .iso is
served as an NetInstall image via isoredirect.

The lack of $version in these two queries was a bug, but it didn't
really affect anything because all the current .isos have a version
string embedded in their names.